requirements.txt
Django>=5.0.2,<5.1
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.9,<3.0
drf-spectacular>=0.27.1,<0.28
Pillow>=10.2.0,<10.3
uwsgi>=2.0.24,<2.1
django-cors-headers>=4.3.1,<4.4
dj-rest-auth>=5.0.2,<5.1
djangorestframework-simplejwt>=5.3.1,<5.4
django-allauth>=0.61.1,<0.62
unstructured>=0.11.8,<0.14.0
langchain>=0.1.14,<0.2.0
langchain-openai>=0.1.1,<0.2.0
selenium>=4.19.0,<4.20.0
beautifulsoup4<=4.12.3,<4.13.0